prompt_generator_for_content_transformation_cases/requirements.txt
fastapi
uvicorn
python-dotenv
openai
termcolor
python-multipart
jinja2
aiofiles 